Move files around. Now have a proper library, and the unit tests are in sub-libraries.
