+++ go/cmd/vtgateclienttest/goclienttest/main.go
@@ -2,13 +2,14 @@
 // Use of this source code is governed by a BSD-style
 // license that can be found in the LICENSE file.
 
-package main
+package goclienttest
 
 import (
 	"encoding/json"
 	"testing"
 	"time"
 
+	"github.com/youtube/vitess/go/cmd/vtgateclienttest/services"
 	"github.com/youtube/vitess/go/vt/callerid"
 	"github.com/youtube/vitess/go/vt/topo"
 	"github.com/youtube/vitess/go/vt/vtgate/vtgateconn"
@@ -36,7 +37,7 @@ func testCallerID(t *testing.T, conn *vtgateconn.VTGateConn) {
 		t.Errorf("failed to marshal callerid: %v", err)
 		return
 	}
-	query := callerIDPrefix + string(data)
+	query := services.CallerIDPrefix + string(data)
 
 	// test Execute forwards the callerID
 	if _, err := conn.Execute(ctx, query, nil, topo.TYPE_MASTER); err != nil {
@@ -46,7 +47,8 @@ func testCallerID(t *testing.T, conn *vtgateconn.VTGateConn) {
 	// FIXME(alainjobart) add all function calls
 }
 
-func testGoClient(t *testing.T, protocol, addr string) {
+// TestGoClient runs the test suite for the provided client
+func TestGoClient(t *testing.T, protocol, addr string) {
 	// Create a client connecting to the server
 	ctx := context.Background()
 	conn, err := vtgateconn.DialProtocol(ctx, protocol, addr, 30*time.Second)
