+++ go/cmd/vtgateclienttest/gorpcclienttest/gorpc_goclient_test.go
@@ -2,13 +2,15 @@
 // Use of this source code is governed by a BSD-style
 // license that can be found in the LICENSE file.
 
-package main
+package gorpcclienttest
 
 import (
 	"net"
 	"net/http"
 	"testing"
 
+	"github.com/youtube/vitess/go/cmd/vtgateclienttest/goclienttest"
+	"github.com/youtube/vitess/go/cmd/vtgateclienttest/services"
 	"github.com/youtube/vitess/go/rpcplus"
 	"github.com/youtube/vitess/go/rpcwrap/bsonrpc"
 	"github.com/youtube/vitess/go/vt/vtgate/gorpcvtgateservice"
@@ -19,7 +21,7 @@ import (
 
 // TestGoRPCGoClient tests the go client using goRPC
 func TestGoRPCGoClient(t *testing.T) {
-	service := createService()
+	service := services.CreateServices()
 
 	// listen on a random port
 	listener, err := net.Listen("tcp", ":0")
@@ -41,5 +43,5 @@ func TestGoRPCGoClient(t *testing.T) {
 	go httpServer.Serve(listener)
 
 	// and run the test suite
-	testGoClient(t, "gorpc", listener.Addr().String())
+	goclienttest.TestGoClient(t, "gorpc", listener.Addr().String())
 }
